+++ swarm/CHANGELOG.md
@@ -10,8 +10,11 @@
   Individual protocols should not keep connections alive for longer than necessary.
   Users should use `swarm::Config::idle_connection_timeout` instead.
   See [PR 4656](https://github.com/libp2p/rust-libp2p/pull/4656).
+- Remove `KeepAlive::Until`.
+  See [PR 4677]
 
 [PR 4120]: https://github.com/libp2p/rust-libp2p/pull/4120
+[PR 4677]: https://github.com/libp2p/rust-libp2p/pull/4677
 
 ## 0.43.5