+++ swarm/CHANGELOG.md
@@ -10,6 +10,8 @@
   Individual protocols should not keep connections alive for longer than necessary.
   Users should use `swarm::Config::idle_connection_timeout` instead.
   See [PR 4656](https://github.com/libp2p/rust-libp2p/pull/4656).
+- Deprecate `keep_alive_timeout` in `OneShotHandlerConfig`.
+  See [PR 4680](https://github.com/libp2p/rust-libp2p/pull/4680).
 
 [PR 4120]: https://github.com/libp2p/rust-libp2p/pull/4120